requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Pillow>=10.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
opencv-python>=4.9.0
python-dotenv>=1.0.0
joblib>=1.3.0
